+++ core/src/com/unciv/ui/components/widgets/ExpanderTab.kt
@@ -1,4 +1,4 @@
-package com.unciv.ui.components
+package com.unciv.ui.components.widgets
 
 import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.math.Interpolation
