+++ core/src/com/unciv/ui/popups/options/KeyBindingsTab.kt
@@ -5,7 +5,7 @@ import com.unciv.GUI
 import com.unciv.UncivGame
 import com.unciv.models.ruleset.RulesetCache
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.ExpanderTab
+import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.KeyCapturingButton
 import com.unciv.ui.components.TabbedPager
 import com.unciv.ui.components.extensions.toLabel
