+++ core/src/com/unciv/ui/popups/options/ModCheckTab.kt
@@ -13,7 +13,7 @@ import com.unciv.models.ruleset.validation.RulesetError
 import com.unciv.models.ruleset.validation.RulesetErrorSeverity
 import com.unciv.models.ruleset.validation.UniqueValidator
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.ExpanderTab
+import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.TabbedPager
 import com.unciv.ui.components.TranslatedSelectBox
 import com.unciv.ui.components.extensions.surroundWithCircle
