+++ core/src/com/unciv/ui/screens/diplomacyscreen/OffersListScroll.kt
@@ -22,7 +22,7 @@ import com.unciv.logic.trade.TradeType.WarDeclaration
 import com.unciv.logic.trade.TradeType.values
 import com.unciv.models.ruleset.tile.ResourceSupplyList
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.ExpanderTab
+import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.extensions.disable
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.images.IconTextButton
