+++ core/src/com/unciv/ui/screens/overviewscreen/UnitOverviewTab.kt
@@ -16,7 +16,7 @@ import com.unciv.logic.map.tile.Tile
 import com.unciv.models.UnitActionType
 import com.unciv.models.UpgradeUnitAction
 import com.unciv.models.ruleset.unit.BaseUnit
-import com.unciv.ui.components.ExpanderTab
+import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.Fonts
 import com.unciv.ui.components.TabbedPager
 import com.unciv.ui.components.UnitGroup
