+++ core/src/com/unciv/ui/screens/worldscreen/mainmenu/WorldScreenMusicPopup.kt
@@ -9,7 +9,7 @@ import com.unciv.UncivGame
 import com.unciv.models.metadata.GameSettings
 import com.unciv.models.metadata.ScreenSize
 import com.unciv.ui.audio.MusicController
-import com.unciv.ui.components.ExpanderTab
+import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.Fonts
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.components.extensions.setSize
